// frontend/src/sections/accountdetails/connectors/types/types.ts
interface Connector {
  _key?: string;
  name: string;
  type: string;
  appGroup: string;
  appGroupId: string;
  authType: string;
  appDescription: string;
  appCategories: string[];
  iconPath: string;
  isActive: boolean;
  isConfigured: boolean;
  supportsRealtime: boolean;
  createdAtTimestamp: number;
  updatedAtTimestamp: number;
}